go-to-gym-platform/requirements.txt
Django>=5.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
